requirements.txt

fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pandas>=2.0.0
chromadb>=0.4.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
pillow